Extract dynamic admission control plane
