+++ tools/enforcement-fixtures/overlays/layering-tree-in-driver-module/crates/shelterwood/src/driver/admission_control.rs
@@ -1,0 +1,5 @@
+// Known-bad fixture: nested driver modules remain part of the driver layer,
+// so an upward tree reference here must be flagged by the layering check.
+pub fn shell() {
+    tree::Lowered::admit();
+}
